fixing testing issues and reinstalling mocha
