+++ package.json
@@ -1,21 +1,31 @@
 {
   "name": "webrtc-signal-http",
-  "version": "1.8.0",
+  "version": "1.8.3",
   "description": "opinionated webrtc signal provider using http as a protocol",
-  "main": "lib/index.js",
+  "main": "dist/lib/index.js",
   "directories": {
     "lib": "lib",
     "test": "test"
   },
   "bin": {
-    "webrtc-signal-http": "./dist/index.js"
+    "webrtc-signal-http": "./dist"
   },
+  "files": [
+    "lib/**/*",
+    "package-lock.json",
+    "index.ts",
+    "tsconfig.json",
+    "LICENSE",
+    "dist/**/*"
+  ],
   "scripts": {
     "test": "ts-mocha test/**/*.ts",
+    "prepare": "npm run build",
     "build": "tsc -p .",
     "lint": "tslint -c tslint.json -p tsconfig.json --fix",
     "start": "node dist/index.js"
   },
+  "types": "index.d.ts",
   "keywords": [
     "webrtc",
     "signal"
@@ -32,6 +42,7 @@
     "@types/cors": "^2.8.4",
     "@types/express": "^4.16.2",
     "@types/express-bunyan-logger": "^1.3.1",
+    "@types/mocha": "^5.0.0",
     "mocha": "^5.0.0",
     "supertest": "^3.0.0",
     "ts-mocha": "^6.0.0",
