+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
     "compilerOptions": {
       "module": "commonjs",
+      "declaration": true,
       "target": "es5",
       "noImplicitAny": true,
       "sourceMap": true,
